# apps/broker-api/src/server.ts
/**
 * @fileoverview Thought Broker API Server
 * RESTful API for Mobius Systems deliberation and consensus
 */
import path from 'path';
import express, { Request, Response } from 'express';
import cors from 'cors';
import helmet from 'helmet';
import { WebSocketServer } from 'ws';
import http from 'http';
import dotenv from 'dotenv';
import swaggerUi from 'swagger-ui-express';
import YAML from 'yamljs';
import { DeliberationEngine } from './deliberation/engine';
import { DeliberationRequestSchema, APIResponse, WSEvent } from './types';
import {
  createTrialSessionHandler,
  recordTrialEventHandler,
  closeTrialSessionHandler,
} from './routes/trials';
import {
  getTrialSummaryHandler,
  getTrialEventsHandler,
  getGlobalTrialStatsHandler,
  listTrialSummariesHandler,
} from './routes/trialAnalytics';
import { getMicSuggestionForTrialHandler } from './routes/mic';
import { exportKtt001Handler } from './routes/exportKtt';
import {
  initTrialHandler,
  recruitTrialHandler,
  testTrialHandler,
  getTrialDataHandler,
  analyzeTrialHandler,
  getTrialStatusHandler,
} from './routes/trialManagement';
import { createPublishRouter } from './routes/publish';
import memtRouter from './routes/memt';
import encyclopediaRouter from './routes/encyclopedia';

// Load environment
dotenv.config();

const app = express();
const PORT = Number(process.env.PORT ?? '4005');
const WS_PORT = Number(process.env.WS_PORT ?? '4006');

// Middleware
app.use(helmet({
  contentSecurityPolicy: false, // Allow Swagger UI to load
}));
app.use(cors());
app.use(express.json());

// Swagger UI
try {
  const swaggerDocument = YAML.load(path.join(__dirname, '../openapi.yaml'));
  app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument, {
    customCss: '.swagger-ui .topbar { display: none }',
    customSiteTitle: 'Thought Broker API Documentation',
    customfavIcon: '/favicon.ico',
  }));
  console.log('📚 Swagger UI available at /api-docs');
} catch (error) {
  console.warn('⚠️  Swagger UI setup failed:', error);
}

app.use(
  express.static(path.join(__dirname, '../public'), {
    index: false,
  }),
);

app.get('/dashboard/ktt-001', (_req, res) => {
  res.sendFile(path.join(__dirname, '../public/ktt-001-dashboard.html'));
});

// Initialize Deliberation Engine
const engine = new DeliberationEngine();

// ============================================================================
// REST API ENDPOINTS
// ============================================================================

/**
 * Health check
 */
app.get('/healthz', (req: Request, res: Response) => {
  res.json({
    status: 'healthy',
    service: 'thought-broker-api',
    version: '0.1.0',
    timestamp: new Date()
  });
});

/**
 * Thought Broker specific health
 */
app.get('/v1/loop/health', (req: Request, res: Response) => {
  const activeDeliberations = engine.listDeliberations();
  
  res.json({
    status: 'healthy',
    service: 'thought-broker-loop',
    activeDeliberations: activeDeliberations.length,
    timestamp: new Date()
  });
});

/**
 * POST /v1/deliberate - Initiate new deliberation
 */
app.post('/v1/deliberate', async (req: Request, res: Response) => {
  try {
    // Validate request
    const validationResult = DeliberationRequestSchema.safeParse(req.body);
    
    if (!validationResult.success) {
      const response: APIResponse<null> = {
        success: false,
        error: {
          code: 'VALIDATION_ERROR',
          message: 'Invalid deliberation request',
          details: validationResult.error.issues
        },
        timestamp: new Date()
      };
      return res.status(400).json(response);
    }

    // Start deliberation
    const deliberationId = await engine.deliberate(validationResult.data);
    const response: APIResponse<{ deliberationId: string }> = {
      success: true,
      data: { deliberationId },
      timestamp: new Date()
    };
    res.status(202).json(response);
  } catch (error) {
    console.error('Error initiating deliberation:', error);
    
    const response: APIResponse<null> = {
      success: false,
      error: {
        code: 'INTERNAL_ERROR',
        message: error instanceof Error ? error.message : 'Unknown error'
      },
      timestamp: new Date()
    };
    
    res.status(500).json(response);
  }
});

/**
 * GET /v1/deliberation/:id - Get deliberation status
 */
app.get('/v1/deliberation/:id', (req: Request, res: Response) => {
  const { id } = req.params;
  const deliberation = engine.getDeliberation(id);
  if (!deliberation) {
    const response: APIResponse<null> = {
      success: false,
      error: {
        code: 'NOT_FOUND',
        message: 'Deliberation not found'
      },
      timestamp: new Date()
    };
    return res.status(404).json(response);
  }

  const response: APIResponse<typeof deliberation> = {
    success: true,
    data: deliberation,
    timestamp: new Date()
  };
  res.json(response);
});

// ============================================================================
// MEMT ROUTING (Multi-Engine Model Taxonomy)
// ============================================================================
app.use('/v1/memt', memtRouter);

// ============================================================================
// ENCYCLOPEDIA ROUTES
// ============================================================================
app.use('/v1/encyclopedia', encyclopediaRouter);

// ============================================================================
// INTEGRITY TIER PUBLISHING
// ============================================================================
app.use(createPublishRouter());

/**
 * GET /v1/deliberations - List all deliberations
 */
app.get('/v1/deliberations', (req: Request, res: Response) => {
  const deliberations = engine.listDeliberations();
  
  const response: APIResponse<typeof deliberations> = {
    success: true,
    data: deliberations,
    timestamp: new Date()
  };
  res.json(response);
});

/**
 * POST /v1/grade - Grade MII without full deliberation
 */
app.post('/v1/grade', async (req: Request, res: Response) => {
  try {
    const { MIIGrader } = await import('./mii/grader');
    const grader = new MIIGrader();
    
    const result = await grader.grade(req.body);
    const response: APIResponse<typeof result> = {
      success: true,
      data: result,
      timestamp: new Date()
    };
    res.json(response);
  } catch (error) {
    console.error('Error grading MII:', error);
    
    const response: APIResponse<null> = {
      success: false,
      error: {
        code: 'GRADING_ERROR',
        message: error instanceof Error ? error.message : 'Unknown error'
      },
      timestamp: new Date()
    };
    
    res.status(500).json(response);
  }
});

// ============================================================================
// KTT TRIAL-001 API
// ============================================================================
app.post('/v1/trials/ktt-001/session', createTrialSessionHandler);
app.post('/v1/trials/ktt-001/event', recordTrialEventHandler);
app.post('/v1/trials/ktt-001/close', closeTrialSessionHandler);

app.get('/v1/trials/ktt-001', listTrialSummariesHandler);
app.get('/v1/trials/ktt-001/stats', getGlobalTrialStatsHandler);
app.get('/v1/trials/ktt-001/export', exportKtt001Handler);
app.get('/v1/trials/ktt-001/:trialId/summary', getTrialSummaryHandler);
app.get('/v1/trials/ktt-001/:trialId/events', getTrialEventsHandler);
app.get('/v1/trials/ktt-001/:trialId/mic-suggestion', getMicSuggestionForTrialHandler);

// ============================================================================
// TRIAL MANAGEMENT API (KTT Trial-001 Extended)
// ============================================================================
app.post('/v1/trials/init', initTrialHandler);
app.get('/v1/trials/:trialId', getTrialStatusHandler);
app.post('/v1/trials/:trialId/recruit', recruitTrialHandler);
app.post('/v1/trials/:trialId/test', testTrialHandler);
app.get('/v1/trials/:trialId/data', getTrialDataHandler);
app.post('/v1/trials/:trialId/analyze', analyzeTrialHandler);

// ============================================================================
// WEBSOCKET SERVER
// ============================================================================

const server = http.createServer(app);
let wss: WebSocketServer | null = null;

// Track connected clients
const clients = new Set<any>();

function startWebSocketServer(): void {
  wss = new WebSocketServer({ port: WS_PORT });

  wss.on('connection', (ws) => {
    console.log('[WebSocket] Client connected');
    clients.add(ws);
    ws.on('close', () => {
      console.log('[WebSocket] Client disconnected');
      clients.delete(ws);
    });
    ws.on('error', (error) => {
      console.error('[WebSocket] Error:', error);
      clients.delete(ws);
    });
  });
}

// Forward engine events to WebSocket clients
engine.on('event', (event: WSEvent) => {
  const message = JSON.stringify(event);
  
  clients.forEach((client) => {
    if (client.readyState === 1) { // OPEN
      client.send(message);
    }
  });
});

// ============================================================================
// START SERVER
// ============================================================================

if (require.main === module) {
  startWebSocketServer();
  server.listen(PORT, () => {
    console.log(`
╔═══════════════════════════════════════════════════════════╗
║       THOUGHT BROKER API — MOBIUS SYSTEMS v0.1            ║
╚═══════════════════════════════════════════════════════════╝
🔴 DELIBERATION ENGINE: ONLINE
📡 REST API: http://localhost:${PORT}
🔌 WebSocket: ws://localhost:${WS_PORT}

Endpoints:
  GET  /healthz
  GET  /v1/loop/health
  POST /v1/deliberate
  GET  /v1/deliberation/:id
  GET  /v1/deliberations
  POST /v1/grade
  GET  /api-docs          — Swagger UI Documentation

MEMT Endpoints:
  POST /v1/memt/deliberate   — MEMT-routed deliberation
  POST /v1/memt/classify     — Classify task without routing
  GET  /v1/memt/engines      — List engine profiles
  GET  /v1/memt/engine/:id   — Get engine profile
  GET  /v1/memt/health       — MEMT health check

"The Broker is not a service. The Broker is where the system thinks."

— ATLAS, C-132

Ready for consensus...
    `);
  });

  // Graceful shutdown
  process.on('SIGTERM', () => {
    console.log('[Shutdown] SIGTERM received, closing servers...');
    server.close(() => {
      if (wss) {
        wss.close(() => {
          console.log('[Shutdown] Servers closed');
          process.exit(0);
        });
      } else {
        console.log('[Shutdown] Server closed');
        process.exit(0);
      }
    });
  });
}

export default app;